Move emoji_support code to emojies.py.
